move png utilities to age_common
