+++ src/CMakeLists.txt
@@ -53,6 +53,7 @@ endif ()
 find_package(PNG QUIET)
 if (PNG_FOUND)
     message("PNG found => configuring age_tester")
+    add_compile_definitions(LIBPNG_FOUND)
     add_subdirectory(age_tester)
 else ()
     message("PNG not found => age_tester not available")
